shortcuts: make various application shortcuts rebindable
This creates a shortcut object (CapsuleShortcuts) which wraps GSettings
(org.gnome.Capsule{,.Devel}.Shortcuts) at the GSettings path
/org/gnome/Capsule/{,Devel}/Shortcts/ with the option at a future date to
rebind that to a new path for per-profile overrides should it ever be
necessary.

The CapsuleShortcuts object also knows how to rewrite GMenu items so that
we can update the GMenu in .ui files as the shortcuts are changed. This
helps ensure that the user sees the active shortcut for the action. The
cross-reference uses "id" as an attribute to locate the accel.

Shortcuts have a property in CapsuleShortcut which maps to the gschema
entry under the same name.

An XMACROS file is used to define the shortcuts for the C code to make that
a lot easier to manage since we already need to update enough external
places beyond this.

Some work is still necessary for tab/terminal menus.