help-overlay: use shortcuts page of preferences
Instead of using the old GtkShortcutsWindow I wrote for Builder years ago,
just take us to the shortcuts section of the preferences. That has both
the actual shortcut accelerators along with descriptions. If we used the
help-overlay.ui, we wouldn't even be able to edit them. This allows both
searching and editing.