process: abstract GSubprocess to CapsuleProcess
We want to eventually be able to support things that will cross the PID
namespace barrier. Doing so will require solving problems differently based
on the container technology.

This abstracts things into the more simple form of what we need to help
make that possible going forward.

It is expected that the CapsuleContainer may create a different process
subclass specifically for their use and override the virtual methods.