tab-monitor: start on CapsuleTabMonitor for leader tracking
The goal here is to track the leader process for the PTY in a fashion that
is much less overhead than what Console does. For example, we don't need
a whole process tree to peek at the foreground process (when it is in our
own PID namespace).

This uses an exponential backoff up to the max to reduce how often we poll
and also encourages the other tabs to align their wakeups to the same time
by bracketing to the whole seconds of the monotonic clock.

The actual work to poll is still done by CapsuleProcess which could have
different implementations by container. But the tracking for changes is
done by the process-leader-kind property of CapsuleTabMonitor.