tab-monitor: improve polling interactivity
We can take some shortcuts to poll on a lower latency interval for the
first poll based on the input key. For example, if we get ctrl+d, we might
be leaving sudo and can update that state quicker. Additionally, if we hit
Enter/Return we might be entering a command and that additional state could
mean we get the appropriate color bar sooner.