css: tweak padding for the terminal
Okay, so what I really don't like is that we have a gap between the upper
padding area and the topbar on scrollback. But that is because we're
getting clipping from VteTerminal which we *might* be able to hack around
instead.

But either way, having those at 0 gives this weird feel the rest of the
time when we're not in scrollback. So do this for now until we can find a
better way to hack around the scrollback thing.