terminal: make scrolling beautiful in presence of padding
We have padding at the edges of our terminal widget so that it both looks
nice and avoids the rounded corners clipping various content as you type.

However, that results in a clip node during scrolling which prevents
terminal contents from reaching the top/bottom of the scrolledwindow.

Obviously that gap that creates is non-ideal, so this hack removes the
clip node (in favor of our already established parent clip node) so that
we can draw to the boundaries cleanly.

Now we can have both beautiful scrolling and beautiful padding.