container-provider: implement container merge
This will take an array of containers that were polled by the container
provider and merge them into the GListModel based on container id.

Containers not found in the new list will be removed. Containers not found
in the original set will be added.

This is convenience API so that tooling like PromptPodmanContainer can be
relatively state agnostic in terms of polling. Just part the entire data
set and let the base class deal with merging.