agent: start on agent implementation
This is meant to be run on the host (possibly escaping the flatpak) to
manage how we create/process containers. The goal here will be to proxy
work to that process from our UI process.

It will be able to manage dealing with container technologies directly
so that we can keep that knowledge out of the UI process.