app: move to using client for process management
This is a fairly large change but one I've wanted from the start. We need
to push the control of most things to an agent we control on the host.

This allows us to create PTY devices where we want, as well as spawn
various containers without having to first jump through the hoop to the
host. We just talk to our agent which is running there, which is similar
in what happens but with more ease.