process: use proper PTY fd for checking foreground/leader
This ensures that we use the PTY master for API like tcgetpgrp() because
we can't use the FD that is mapped into the child. This was fallout from
changing things to pass the PTY to be mapped in to the agent.

Theoretically we could drop some roundtrips there, but I don't think it's
really worth it since we have a private socketpair to the agent.