flatpak: include patches necessary for precmd/preexec/notification
These are currently part of Fedora but we want to be able to use them too.
Additionally, we make them available to the GTK 4 port as well as the
GTK 3 version. Fedora only needed the GTK 3 version for gnome-terminal but
we want them too.

Ideally, we'll get these upstream instead.