window: check for current container (requires VTE patches)
This requires the VTE patches that are part of the Flatpak build. But the
goal here is to track the container appropriately so that we can know what
container has been entered.

Note that currently because of how toolbox is designed, it only emits
the appropriate escape sequences if you're running a machine with
/etc/os-release containing ID=fedora and VARIANT=workstation or
VARIANT=silverblue. This is not idea and can hopefully be fixed in the
future.

This omits many Fedora users including those on Asahi where we have
ID_LIKE=fedora but ID=fedora-asahi-remix.