window: add containers to split button menu
This needs iteration badly, and the action is not wired up. But lets start
by setting the containers somewhere that is visible where we can figure out
what the right action is.

One possible action is that we take the current profile, but the activated
container to setup the new tab/window.