profile: add preserve-container profile option
This adds a preference on the profile object that specifies if the current
container should be applied to new terminal tabs/windows instead of the
default-container for the profile.

This is useful for situations where you have a profile you do not want to
take the container with you. Some people prefer this to the default of
always copying the current container, such as gnome-terminal on Fedora.