tab: disable icon for containers
I don't know yet what we should be doing here for this. For example, we
might want to look at /etc/os-release in the conatiner to get an icon for
the distribution. Or we might just want to say "toolbox-symbolic" or
something like that.

Or, just show nothing, which is probably the best option currently.