Flatpak: Use type dir for the repository module
Otherwise, git default at looking and building the "master" branch
or the specified one, but we want it to be building the checkout
instead.

gnome-builder and our CI do that already, but flatpak-builder falls
on it's face as it will try to build the module we specified.

```
fatal: ambiguous argument 'master': unknown revision or path not in the working tree.
Use '--' to separate paths from revisions, like this:
'git <command> [<revision>...] -- [<file>...]'
Failed to download sources: module prompt: Child process exited with code 128
```