menus: use separate section for "My Computer"
It's not a container so it shouldn't be listed under them. We also don't
need to show it when it's the only option. This uses an extra action to
get that behavior and uses the profiles/containers list model to enforce
the widget visibility (intead of the GMenu like we did before, as that is
more problematic).