session: add windows/tabs to session state
We still need to find the right place to call into this stuff where we
will be able to still get the windows before they're destroyed. We also
would like to find a trigger from the session to determine when shutdown
is happening otherwise we would need a quit action to do this.