window: fix tab container inheritance
If we created a new tab using the menu, we know what container it was when
we created it. Podman won't tell us what container it was through escape
sequences, so we need to do this manually using the same thing we started
with or we risk just using the "profile default" which could be different.