app: try to canonicalize relative paths
This one relies on the other side actually sending the appropriate CWD
but we can expected that from GLib at least so should be fine.

But if we get something relative, try to resolve it from the CWD of the
call site rather than whatever the single-instance application is started
from (which is like $HOME).

Related #26