find-bar: fix searching outside current page
This points out something I still don't like though with our find-bar. It
is a single instance of the widget (because why waste extra widgets) but
it won't preserve state as you move between them.

Either way, doing what this patch does (by only updating the regex when
the entry changes) is better than how it was broken before.

Fixes #39