session: improve unmaximized size when restoring session state
It would be really nice if we had a value of the column/row size before
the window went maximized, but there are so many ways that could break
such as zoom changes or client events.

So just do the minimum necessary to be less annoying when leaving the
maximized state after restoring session.

Fixes #45