agent: add API for URI translations
Currently we just have passthrough for the session container. However the
goal here is to undo the overlay layers/image/etc from a podman container
so that we can get to the real file underneath.

We need to do that because we must be able to get access to the file from
within the Flatpak UI process to be able to open the URI on the host
system. That is somewhat problematic because we may not always be able to
resolve it. But such is life for now.

Another option is to use the container to `cat` the contents of a URI into
a FD we control, but that is ripe for disaster.

We will also have to open up filesystem for home/host so that we can try
to get access to files at these particular paths.