tab: move uri opening to tab
We are going to have to cook a bunch of stuff here through the container
IPC to the agent so that we might get a chance at resolving the URI from
the Flatpak mount namespace.

So might as well just bring over the libportal bits too so we can put up
a toast if it fails.