window: special case size for new-window when maximized
If we are maximized, making another maximized window on top can be a bit
annoying. My guess is that the most likely thing in this scenario is that
you want to either discard the window (transiently) or move it elsewhere.

Either of which can be slightly annoying if we stay maxmized.