agent: translate URIs for podman containers
This currently does not work because the linchpin is podman not injecting
an extra PTY into the situation. That is currently a problem because we
need to get the real foreground process to use it's mount namespace to
translate paths after reconstructing it's mount namespace (then translating
it to something we can see from the host).

Once that is fixed, most of this should start working like we do for ELF
translation in Sysprof.