app: fallback to getent to discover user shell
If we fail to locate the user's preferred shell from the host, try to get
the user's default shell from inside the container.

This works around an issue where `which` is unavailable inside a container
(arch:latest for example) and we cannot discover a shell to execute even
though /usr/bin/bash is there.

Fixes #84