a11y: make terminal accessible with GtkAccessibleText
This is a very minimal first step at making the terminal accessible to
screen readers. It's currently guarded behind a preference as I imagine
it could have an adverse affect on performance of both Ptyxis as well as
the screen reader given the current design of GtkAccessibleText.

It isn't exact as what is in GTK 3, as that will need more work in GTK to
support things read like "backspace x".