agent: get proxy settings from ptyxis-agent
We cannot rely on proxy settings inside the Flatpak because they will not
reflect what is set on the host.

This tries to access the host proxy settings and fallback to whatever
environment we can copy in the case we're not on GNOME nor the desktop
schemas from GNOME are available.