parking-lot: force-quit tab when removing
If we don't force the process to exit when removing the tab, then there is
a chance we'll never break our reference cycles causing the tab+process to
leak until something terminates the process.

This could happen when closing a tab using the close button instead of
exiting the shell with `exit` or ctrl+d.

Fixes #142