client: work around misconfigured passwd in containers
If a container says the user's preferred shell is `/bin/zsh` but `/bin/zsh`
does not exist, then we may end up trying to execute a non-existant shell
via `sh -c $(getent passwd $(whoami) | cut -d : -f 7)` which can leave a
user unable to get into their container.

This works around that by trying to exec the appropriate shell replacing
the invoked `sh`. If the shell specified in passwd is executable, we use
that, otherwise fallback to `sh`.

Fixes: #174