app: Save and restore window state via Wayland session management
Save the session ID obtained through GtkApplication in the GVariant
file saved in disk, and give a name to each of the windows saved
there.

In the next run(s), we can apply the GtkApplication session identifier
from the prior run, and let the compositor restore window states through
their identifiers.