gnome.palette: match GNOME Terminal on palette colors
This updates the GNOME palette to match the colors from GNOME Terminal and
what was suggested for GNOME/console!80.

This also ensures that we have more contrast (white on black, black on
white) for the foreground/background.

To keep the GNOME-styling for the titlebars, those colors are specified
manually so that overview operations also have appropriate contrast
between the terminal background and the window background.

Related: #191