Don't override the inspector
By forcing MAXINT as the theme priority, we guarantee that this theme will be used absolutely and under any circumstances, even if the inspector would like to do something different.
So don't do that.

We do want to override any user settings from other config methods, but this is still a user setting. So pick something related to user settings.