GNOME Terminal.palette: add light palette variant
This uses the same color palette but tweaks the foreground/background and
cursor to match what GNOME Terminal has. This allows it to follow the
system light/dark style.

We can probably do better by adding an option to inherit the adwaita style
instead of overriding it because the selection colors feel off here. For
that we likely want to just take Adwaita styling.