GNOME Terminal.palette: tweak titlebar for warm undertone
The titlebar had a completely neutral color while the terminal background
has a warm undertone. This tries to balance them a bit so we get a lighter
color without _too_ much of the undertone coming through but still not a
fully neutral. Found from mixing the 3rd tango color (3D3846) with black
and choosing a step relatively close to where we'd want the color to end
up w/ 303030.