app: fix synthesizing of exit code
If we fail to get the process before we could process it, just assume it
was normal exit for now. We guarantee correctness here unless we change
the agent design to provide a reverse proxy object path to deliver the
notification of exit status.

Long term, we need to do that alternate design.

Fixes: #157