agent: avoid races in exit notification for processes
If we race against a process exiting we might not be able to subscribe to
it's exited/signalled signals in time.

This provides that same information on the agent interface directly which
allows the PtyxisApplication to cache the result to be reused.

This is a better fix for the race in #157