Make double-click selection boundary configurable
Use the GSettings key "word-char-exceptions" to control how much text is
selected by a double-click. Because VTE accepts null, use a GSettings
[`ms` type (maybe string)][1]. We'll use a default value of `nothing`,
meaning the default Ptyxis behavior will not change.

Whenever a new VTE is instantiated, or the GSettings value is changed,
the value is read, and passed to the VTE, controlling the double-click
behavior.

[1]: https://docs.gtk.org/glib/gvariant-format-strings.html