window: implement behaviors for middle-mouse-click on tab widget
This implements the functionality for #285 to change the behavior of
middle mouse click on a tab widget.

The default is 'close', which matches the default libadwaita behavior.

Additiionally there is 'paste' which can paste the clipboard contents to
the tab along with raising the tab to foreground.

Thirdly there is 'nothing' which will only raise the tab widget but take
no action on the tab itself.

To set paste on middle-click of the tab widget:

 gsettings set org.gnome.Ptyxis tab-middle-click paste

Closes: #285