settings: add global gsetting to disable osc-title
This is an escape hatch for cases where you want all the osc-title to be
ignored. Since that is generally frowned upon in the VTE space, there is
no UI action for this and only the GSetting.

I could have gone the route of making this per-profile, but again, this is
an escape hatch so I'd rather not do it at that level of granularity.